fixing elo type in playergame

+++ BBStats/Data/Entites/PlayerGame.cs
@@ -4,7 +4,7 @@ public class PlayerGame
 {
 	public int GameId { get; set; }
 	public Int64 PlayerId { get; set; }
-	public int EloBefore { get; set; }
-	public int EloAfter { get; set; }
+	public double EloBefore { get; set; }
+	public double EloAfter { get; set; }
 	public Game Game { get; set; }
 }
